requirements.txt
crawl4ai
python-docx
